+++ kadence/.eslintrc.json
@@ -11,7 +11,8 @@
     "eslint:recommended",
     "airbnb-base",
     "prettier",
-    "next/core-web-vitals"
+    "next/core-web-vitals",
+    "next/babel"
   ],
   "settings": {
     "import/resolver": {
