+++ kadence/pages/api/devices/getDevices.js
@@ -6,7 +6,6 @@ const handler = nextConnect();
 handler.use(middleware);
 
 handler.get(async (req, res) => {
-
     let doc = await req.db.collection('Devices').findOne()
     console.log(doc);
     res.json(doc);
