+++ kadence/pages/api/music/getPlatform.js
@@ -6,7 +6,6 @@ const handler = nextConnect();
 handler.use(middleware);
 
 handler.get(async (req, res) => {
-
     let doc = await req.db.collection('Music').findOne()
     console.log(doc);
     res.json(doc);
